Implemented ElevationMap.java and testing for this class
