+++ .idea/jarRepositories.xml
@@ -31,5 +31,10 @@
       <option name="name" value="maven" />
       <option name="url" value="https://jitpack.io" />
     </remote-repository>
+    <remote-repository>
+      <option name="id" value="maven2" />
+      <option name="name" value="maven2" />
+      <option name="url" value="https://esri.jfrog.io/artifactory/arcgis" />
+    </remote-repository>
   </component>
 </project>
